Separate make commands that require sudo
