+++ .github/workflows/ci-cd.yml
@@ -58,9 +58,11 @@ jobs:
           echo "Building for $OS:$ARCH"
           cd $GITHUB_WORKSPACE
           if [[ $OS == "linux" && $ARCH == "amd64" ]]; then
-            make OS=$OS ARCH=$ARCH
+            make OS=$OS ARCH=$ARCH modcheck swagger binary binary-minimal binary-debug cli bench exporter-minimal verify-config test
+            sudo env "PATH=$PATH" make privileged-test test-clean
+            make covhtml check
           else
-            make OS=$OS ARCH=$ARCH binary binary-minimal binary-debug cli bench exporter-minimal
+            make OS=$OS ARCH=$ARCH binary binary-minimal binary-debug cli bench exporter-minimal test
           fi
         env:
           S3MOCK_ENDPOINT: localhost:4566
