+++ pkg/cli/client_test.go
@@ -1,5 +1,5 @@
-//go:build extended
-// +build extended
+//go:build extended || needsudo
+// +build extended needsudo
 
 package cli //nolint:testpackage
 
@@ -195,62 +195,6 @@ func TestTLSWithoutAuth(t *testing.T) {
 				panic(err)
 			}
 			defer os.RemoveAll(destCertsDir)
-
-			args := []string{"imagetest"}
-			imageCmd := NewImageCommand(new(searchService))
-			imageBuff := bytes.NewBufferString("")
-			imageCmd.SetOut(imageBuff)
-			imageCmd.SetErr(imageBuff)
-			imageCmd.SetArgs(args)
-			err := imageCmd.Execute()
-			So(err, ShouldBeNil)
-		})
-	})
-
-	Convey("Privileged certs - Make a new controller", t, func() {
-		caCert, err := ioutil.ReadFile(CACert)
-		So(err, ShouldBeNil)
-		caCertPool := x509.NewCertPool()
-		caCertPool.AppendCertsFromPEM(caCert)
-
-		resty.SetTLSClientConfig(&tls.Config{RootCAs: caCertPool, MinVersion: tls.VersionTLS12})
-		defer func() { resty.SetTLSClientConfig(nil) }()
-		conf := config.New()
-		conf.HTTP.Port = SecurePort2
-		conf.HTTP.TLS = &config.TLSConfig{
-			Cert:   ServerCert,
-			Key:    ServerKey,
-			CACert: CACert,
-		}
-
-		ctlr := api.NewController(conf)
-		ctlr.Config.Storage.RootDirectory = t.TempDir()
-		go func() {
-			// this blocks
-			if err := ctlr.Run(context.Background()); err != nil {
-				return
-			}
-		}()
-
-		// wait till ready
-		for {
-			_, err := resty.R().Get(BaseURL2)
-			if err == nil {
-				break
-			}
-			time.Sleep(100 * time.Millisecond)
-		}
-
-		defer func() {
-			ctx := context.Background()
-			_ = ctlr.Server.Shutdown(ctx)
-		}()
-
-		Convey("Certs in privileged path", func() {
-			configPath := makeConfigFile(
-				fmt.Sprintf(`{"configs":[{"_name":"imagetest","url":"%s/v2/_catalog","showspinner":false}]}`,
-					BaseSecureURL2))
-			defer os.Remove(configPath)
 
 			args := []string{"imagetest"}
 			imageCmd := NewImageCommand(new(searchService))
